+++ src/Bical.Api/Responses/BirthNoteDtoResponse.cs
@@ -4,7 +4,7 @@ namespace Bical.Api.Responses
 {
     public class BirthNoteDtoResponse
     {
-        public ushort BirthNoteId { get; set; }
+        public ulong BirthNoteId { get; set; }
         public string DisplayedName { get; set; }
         public DateTime BirthDate { get; set; }
         public DateTime Added { get; set; }
